Issue resolved.
The script checks weather the file that has to be backed up exists or not . If it doesn't exist , it gives an error saying " backup file not found".
If the backup file exists and destination file also exitst , backup is done.
If the backup file exists and destination file doesn't exist , it creates a destination file with the name same as given parameter and backup is done.